+++ src/rpc/net.cpp
@@ -196,7 +196,6 @@ static UniValue getpeerinfo(const Config &config,
             obj.push_back(Pair("inflight", heights));
         }
         obj.push_back(Pair("whitelisted", stats.fWhitelisted));
-        obj.push_back(Pair("cashmagic", stats.fUsesCashMagic));
 
         UniValue sendPerMsgCmd(UniValue::VOBJ);
         for (const mapMsgCmdSize::value_type &i : stats.mapSendBytesPerMsgCmd) {