+++ doc/release-notes.md
@@ -6,3 +6,5 @@ This release includes the following features and fixes:
 
  - Remove the newdaaactivationtime configuration.
  - Do not use the NODE_BITCOIN_CASH service bit for preferencial peering anymore.
+ - Only connect to node using the cash magic.
+ - Remove indicator mentionning if a node uses the cash magic getpeerinfo RPC.